+++ dsrag/database/chunk/sqlite_db.py
@@ -1,222 +1,10 @@
-from abc import ABC, abstractmethod
 import os
 import time
-import pickle
 import sqlite3
-from typing import Any, cast
+from typing import Any
 
-
-class ChunkDB(ABC):
-    subclasses = {}
-
-    def __init_subclass__(cls, **kwargs):
-        super().__init_subclass__(**kwargs)
-        cls.subclasses[cls.__name__] = cls
-
-    def to_dict(self):
-        return {
-            "subclass_name": self.__class__.__name__,
-        }
-
-    @classmethod
-    def from_dict(cls, config) -> "ChunkDB":
-        subclass_name = config.pop(
-            "subclass_name", None
-        )  # Remove subclass_name from config
-        subclass = cls.subclasses.get(subclass_name)
-        if subclass:
-            return subclass(**config)  # Pass the modified config without subclass_name
-        else:
-            raise ValueError(f"Unknown subclass: {subclass_name}")
-
-    @abstractmethod
-    def add_document(self, doc_id: str, chunks: dict[int, dict[str, Any]]) -> None:
-        """
-        Store all chunks for a given document.
-        """
-        pass
-
-    @abstractmethod
-    def remove_document(self, doc_id: str) -> None:
-        """
-        Remove all chunks and metadata associated with a given document ID.
-        """
-        pass
-
-    @abstractmethod
-    def get_chunk_text(self, doc_id: str, chunk_index: int) -> str | None:
-        """
-        Retrieve a specific chunk from a given document ID.
-        """
-        pass
-
-    @abstractmethod
-    def get_document(self, doc_id: str) -> FormattedDocument | None:
-        """
-        Retrieve all chunks from a given document ID.
-        """
-        pass
-
-    @abstractmethod
-    def get_document_title(self, doc_id: str, chunk_index: int) -> str | None:
-        """
-        Retrieve the document title of a specific chunk from a given document ID.
-        """
-        pass
-
-    def get_document_summary(self, doc_id: str, chunk_index: int) -> str | None:
-        """
-        Retrieve the document summary of a specific chunk from a given document ID.
-        """
-        pass
-
-    def get_section_title(self, doc_id: str, chunk_index: int) -> str | None:
-        """
-        Retrieve the section title of a specific chunk from a given document ID.
-        """
-        pass
-
-    def get_section_summary(self, doc_id: str, chunk_index: int) -> str | None:
-        """
-        Retrieve the section summary of a specific chunk from a given document ID.
-        """
-        pass
-
-    @abstractmethod
-    def get_all_doc_ids(self, supp_id: str | None = None) -> list[str]:
-        """
-        Retrieve all document IDs.
-        """
-        pass
-
-    @abstractmethod
-    def delete(self) -> None:
-        """
-        Delete the chunk database.
-        """
-        pass
-
-
-class BasicChunkDB(ChunkDB):
-    """
-    This is a basic implementation of a ChunkDB that stores chunks in a nested dictionary and persists them to disk by pickling the dictionary.
-    """
-
-    def __init__(self, kb_id: str, storage_directory: str = "~/dsRAG") -> None:
-        self.kb_id = kb_id
-        self.storage_directory = os.path.expanduser(
-            storage_directory
-        )  # Expand the user path
-        # Ensure the base directory and the chunk storage directory exist
-        os.makedirs(
-            os.path.join(self.storage_directory, "chunk_storage"), exist_ok=True
-        )
-        self.storage_path = os.path.join(
-            self.storage_directory, "chunk_storage", f"{kb_id}.pkl"
-        )
-        self.load()
-
-    def add_document(self, doc_id: str, chunks: dict[int, dict[str, Any]]) -> None:
-        self.data[doc_id] = chunks
-        self.save()
-
-    def remove_document(self, doc_id: str):
-        self.data.pop(doc_id, None)
-        self.save()
-
-    def get_chunk_text(self, doc_id: str, chunk_index: int) -> str | None:
-        if doc_id in self.data and chunk_index in self.data[doc_id]:
-            return self.data[doc_id][chunk_index]["chunk_text"]
-        return None
-
-    def get_document(
-        self, doc_id: str, include_content: bool = False
-    ) -> FormattedDocument | None:
-        if doc_id in self.data:
-            document = self.data[doc_id]
-            title = cast(str, document[0].get("document_title", ""))
-
-            full_document_string = ""
-            if include_content:
-                # Concatenate the chunks into a single string
-                for chunk_index, chunk in document.items():
-                    # Join each chunk text with a new line character
-                    full_document_string += chunk["chunk_text"] + "\n"
-
-            return FormattedDocument(
-                id=doc_id,
-                title=title,
-                content=full_document_string if include_content else None,
-                summary=cast(str, document[0].get("document_summary", "")),
-                created_on=None,
-            )
-
-        else:
-            return None
-
-    def get_document_title(self, doc_id: str, chunk_index: int) -> str | None:
-        if doc_id in self.data and chunk_index in self.data[doc_id]:
-            if "document_title" in self.data[doc_id][chunk_index]:
-                return self.data[doc_id][chunk_index]["document_title"]
-            else:
-                return None
-        return None
-
-    def get_document_summary(self, doc_id: str, chunk_index: int) -> str | None:
-        if doc_id in self.data and chunk_index in self.data[doc_id]:
-            if "document_summary" in self.data[doc_id][chunk_index]:
-                return self.data[doc_id][chunk_index]["document_summary"]
-            else:
-                return None
-        return None
-
-    def get_section_title(self, doc_id: str, chunk_index: int) -> str | None:
-        if doc_id in self.data and chunk_index in self.data[doc_id]:
-            if "section_title" in self.data[doc_id][chunk_index]:
-                return self.data[doc_id][chunk_index]["section_title"]
-            else:
-                return None
-        return None
-
-    def get_section_summary(self, doc_id: str, chunk_index: int) -> str | None:
-        if doc_id in self.data and chunk_index in self.data[doc_id]:
-            if "section_summary" in self.data[doc_id][chunk_index]:
-                return self.data[doc_id][chunk_index]["section_summary"]
-            else:
-                return None
-        return None
-
-    def get_all_doc_ids(self, supp_id: str | None = None) -> list[str]:
-        doc_ids = list(self.data.keys())
-        if supp_id:
-            doc_ids = [
-                doc_id
-                for doc_id in doc_ids
-                if self.data[doc_id][0].get("supp_id", "") == supp_id
-            ]
-        return doc_ids
-
-    def load(self):
-        try:
-            with open(self.storage_path, "rb") as f:
-                self.data = pickle.load(f)
-        except FileNotFoundError:
-            self.data = {}
-
-    def save(self):
-        with open(self.storage_path, "wb") as f:
-            pickle.dump(self.data, f)
-
-    def delete(self):
-        if os.path.exists(self.storage_path):
-            os.remove(self.storage_path)
-
-    def to_dict(self):
-        return {
-            **super().to_dict(),
-            "kb_id": self.kb_id,
-            "storage_directory": self.storage_directory,
-        }
+from dsrag.database.chunk.db import ChunkDB
+from dsrag.database.chunk.types import FormattedDocument
 
 
 class SQLiteDB(ChunkDB):
@@ -322,7 +110,6 @@ class SQLiteDB(ChunkDB):
                 # Join each chunk text with a new line character
                 full_document_string += result[4] + "\n"
 
-        id = doc_id
         title = results[0][1]
         created_on = results[0][3]
         title = results[0][1]
